refactor: Move pydantic models into models file
